+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Fixes
 
 - Classify MetricKit hangs over 500 ms as errors. (#8948)
+- Add hint parameter to public capture methods on `SentrySDK` (#8943)
 
 ## 9.27.0